Room Guide WPF Csharp
The application allows you to move from one room to another. Every room is unique, has its own decorations and special characteristics. There are temperature inf-on only in outside rooms and inf-on about arms and chairs only in inside rooms respectively. You can change those settings to understand how it works. Pay attention on the syntax: it’s the main part of this application. It takes time to understand what’s going on there though that application is quite simple.